Allow materials to be shared
This is a bit gross, and impacts performance. It does however allow us
to re-use materials in multiple hitables, which we will need for the
box type.

A better approach (I suspect) would be to move the ownership of the
materials out of the hitables entirely, and then have the hitables
refer to the materials, either via a raw pointer or an index into some
buffer somewhere. That should be safe provided that the lifetime of the
materials outlives the lifetime of the hitables, and would allow reuse
of materials without resorting to shared_ptr. That is a bigger change
though, and this is good enough for now.